+++ daemon/container.go
@@ -6,11 +6,11 @@ import (
 	"regexp"
 	"time"
 
+	"github.com/docker/docker/api/errors"
 	containertypes "github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/strslice"
 	"github.com/docker/docker/container"
 	"github.com/docker/docker/daemon/network"
-	"github.com/docker/docker/errors"
 	"github.com/docker/docker/image"
 	"github.com/docker/docker/pkg/signal"
 	"github.com/docker/docker/pkg/system"
