+++ daemon/volumes.go
@@ -7,11 +7,11 @@ import (
 	"path/filepath"
 	"strings"
 
+	dockererrors "github.com/docker/docker/api/errors"
 	"github.com/docker/docker/api/types"
 	containertypes "github.com/docker/docker/api/types/container"
 	mounttypes "github.com/docker/docker/api/types/mount"
 	"github.com/docker/docker/container"
-	dockererrors "github.com/docker/docker/errors"
 	"github.com/docker/docker/volume"
 	"github.com/opencontainers/runc/libcontainer/label"
 )
